feature: implemented /src/redux directory into README.md for better documentation
