+++ client/src/components/admin/CategoriesManagement/Type/TypeList/TypeList.jsx
@@ -1,7 +1,7 @@
 import AdminMessage from '@components/admin/AdminMessage/AdminMessage';
 import Spinner from '@UI/Spinner/Spinner';
 import useTimedMessage from '@hooks/admin/useTimedMessage';
-import { loadTypeDelete, responseTypeDelete } from '@redux/admin/selectors';
+import { loadTypeDelete, responseTypeDelete } from '@redux/selectors';
 import { fetchTypes } from '@redux/params/service';
 import { loadTypes, selectTypes } from '@redux/selectors';
 import { useEffect } from 'react';
