+++ client/src/components/admin/PromocodeManagement/PromocodeEdit/PromocodeEdit.jsx
@@ -2,12 +2,7 @@ import ErrorText from '@components/admin/ErrorText/ErrorText';
 import { yupResolver } from '@hookform/resolvers/yup';
 import useIdFromUrl from '@hooks/useIdFromUrl';
 import { editPromocode, getPromocodeById } from '@redux/admin/promocode/service';
-import {
-  errorPromocodeEdit,
-  loadPromocodeEdit,
-  responsePromocodeById,
-  responsePromocodeEdit,
-} from '@redux/admin/selectors';
+import { errorPromocodeEdit, loadPromocodeEdit, responsePromocodeById, responsePromocodeEdit } from '@redux/selectors';
 import { promocodeSchema } from '@validations/admin/promocodeSchema';
 import { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
