+++ client/src/components/admin/PromocodeManagement/PromocodeSelect/PromocodeSelect.jsx
@@ -2,7 +2,7 @@ import cl from './index.module.scss';
 import { useEffect, useState } from 'react';
 import Arrow from '@assets/svg/Admin/Arrow/Arrow';
 import PromocodeSelectOptions from './PromocodeSelectOptions/PromocodeSelectOptions';
-import { isFetchedAllPromocodes } from '@redux/admin/selectors';
+import { isFetchedAllPromocodes } from '@redux/selectors';
 import { useSelector } from 'react-redux';
 
 const PromocodeSelect = () => {
